Added sessions. Can now not get to amin page without login. Login is saved in session.

+++ Oblig-1-ITPE3200/Controllers/ObligController.cs
@@ -1,8 +1,10 @@
 ﻿using Castle.Core.Logging;
+using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
 using Oblig_1_ITPE3200.DAL;
 using Oblig_1_ITPE3200.Models;
+using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 
@@ -14,6 +16,7 @@ namespace Oblig_1_ITPE3200.Controllers
         private readonly IObligRepository _db;
 
         private ILogger<ObligController> _log;
+        private const string _loggedOn = "loggedOn";
 
         public ObligController(IObligRepository db, ILogger<ObligController> log)
         {
@@ -23,6 +26,10 @@ namespace Oblig_1_ITPE3200.Controllers
 
         public async Task<ActionResult> GetAllDiseases()
         {
+            if (string.IsNullOrEmpty(HttpContext.Session.GetString(_loggedOn)))
+            {
+                return Unauthorized();
+            }
             List<Disease> diseases = await _db.GetAllDiseases();
             return Ok(diseases);
         }
@@ -67,6 +74,10 @@ namespace Oblig_1_ITPE3200.Controllers
 
         public async Task<ActionResult> DeleteDisease(int id)
         {
+            if (string.IsNullOrEmpty(HttpContext.Session.GetString(_loggedOn)))
+            {
+                return Unauthorized();
+            }
             bool b = await _db.DeleteDisease(id);
             
             if (!b)
@@ -79,6 +90,10 @@ namespace Oblig_1_ITPE3200.Controllers
 
         public async Task<ActionResult> ChangeDisease(Disease d, List<Symptom> s)
         {
+            if (string.IsNullOrEmpty(HttpContext.Session.GetString(_loggedOn)))
+            {
+                return Unauthorized();
+            }
             bool b = await _db.ChangeDisease(d, s);
 
             if (!b)
@@ -103,14 +118,60 @@ namespace Oblig_1_ITPE3200.Controllers
 
         public async Task<ActionResult> LogIn (User user)
         {
-            bool b = await _db.LogIn(user);
-
-            if (!b)
+            try
             {
-                _log.LogInformation("Did not find matching user.");
-                return NotFound("Did not find user.");
+                if (ModelState.IsValid)
+                {
+                    bool b = await _db.LogIn(user);
+                    if (!b)
+                    {
+                        _log.LogInformation("Log in failed with user=" + user.Username);
+                        HttpContext.Session.SetString(_loggedOn, "");
+                        return Ok(false);
+                    }
+                    HttpContext.Session.SetString(_loggedOn, "LoggedOn");
+                    return Ok(true);
+                }
+                _log.LogInformation("Something wrong in inputvalidation, user=" + user.Username);
+                return BadRequest("Something wrong in inputvalidation");
             }
-            return Ok(b);
+            catch (Exception e)
+            {
+                _log.LogInformation(e.Message);
+                return BadRequest("Something wrong on server.");
+            }
+        }
+
+        // Should be ActionResult
+        public bool LogOut()
+        {
+            try
+            {
+                HttpContext.Session.SetString(_loggedOn, "");
+                return true;
+            }
+            catch (Exception e)
+            {
+                _log.LogInformation(e.Message);
+                return false;
+            }
+        }
+
+        public ActionResult IsLoggedIn()
+        {
+            try
+            {
+                if (HttpContext.Session.GetString(_loggedOn) == "LoggedOn")
+                {
+                    return Ok(true);
+                }
+                return Ok(false);
+            }
+            catch (Exception e)
+            {
+                _log.LogInformation(e.Message);
+                return BadRequest("Something wrong with server. Try again later.");
+            }
         }
 
     }
